+++ Assets/Scripts/IInstaller/UIInstaller.cs
@@ -23,6 +23,7 @@ namespace ElementalBlacksmithStory.Core
             builder.RegisterComponentInHierarchy<SelectShopAmountView>();
             builder.RegisterComponentInHierarchy<ShopItemGridView>();
             builder.RegisterComponentInHierarchy<ShopMasterView>();
+            builder.RegisterComponentInHierarchy<ShopTabView>();
 
             // Presenter 등록
             builder.RegisterEntryPoint<MoneyPresenter>();
@@ -37,7 +38,8 @@ namespace ElementalBlacksmithStory.Core
             builder.RegisterEntryPoint<NPCDialoguePresenter>().AsSelf();
             builder.RegisterEntryPoint<ShopCartPresenter>().AsSelf();
             builder.RegisterEntryPoint<ShopItemGridPresenter>().AsSelf();
-            builder.RegisterEntryPoint<ShopMasterPresenter>().AsSelf();
+            builder.RegisterEntryPoint<ShopMasterPresenter>();
+            builder.RegisterEntryPoint<ShopTabPresenter>().AsSelf();
         }
     }
 }